feat: move shared data access list
