+++ libs/demo/data-access/cars/src/index.ts
@@ -1,1 +1,2 @@
+export { ListFacade as CarsFacade } from '@demo/shared/data-access/list';
 export * from './lib/demo-data-access-cars.module';
