--- libs/demo/data-access/common/list/README.md
@@ -1,7 +1,0 @@
-# demo-data-access-common-list
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test demo-data-access-common-list` to execute the unit tests.
