+++ libs/demo/data-access/users/src/index.ts
@@ -1,4 +1,4 @@
-export { ListFacade as UsersFacade } from '../../common/list/src/lib/+state/facade';
+export { ListFacade as UsersFacade } from '@demo/shared/data-access/list';
 export * from './lib/+state/user/user.facade';
 export * from './lib/demo-data-access-users.module';
 export * from './lib/models/users.model';
