+++ libs/shared/data-access/list/README.md
@@ -1,0 +1,7 @@
+# shared-data-access-list
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-data-access-list` to execute the unit tests.
