+++ libs/shared/data-access/list/project.json
@@ -1,16 +1,16 @@
 {
-  "name": "demo-data-access-common-list",
-  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/demo/data-access/common/list/src",
+  "name": "shared-data-access-list",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/shared/data-access/list/src",
   "prefix": "demo",
-  "tags": [],
+  "tags": ["scope:shared", "type:data-access"],
   "projectType": "library",
   "targets": {
     "test": {
       "executor": "@nrwl/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "libs/demo/data-access/common/list/jest.config.ts",
+        "jestConfig": "libs/shared/data-access/list/jest.config.ts",
         "passWithNoTests": true
       },
       "configurations": {
@@ -24,7 +24,7 @@
       "executor": "@nrwl/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["libs/demo/data-access/common/list/**/*.ts", "libs/demo/data-access/common/list/**/*.html"]
+        "lintFilePatterns": ["libs/shared/data-access/list/**/*.ts", "libs/shared/data-access/list/**/*.html"]
       }
     }
   }
