+++ libs/shared/data-access/list/src/lib/+state/effects.ts
