+++ libs/shared/data-access/list/src/lib/+state/state.service.ts
