+++ libs/shared/data-access/list/src/lib/demo-data-access-common-list.module.ts
@@ -1,4 +1,4 @@
-import { ModuleWithProviders, NgModule, Type } from '@angular/core';
+import { InjectionToken, ModuleWithProviders, NgModule, Type } from '@angular/core';
 import { SharedUtilNotificationModule } from '@demo/shared/util-notification';
 import { LIST_FEATURE_KEY, ListService } from '@ngdux/list';
 import { EffectsModule } from '@ngrx/effects';
@@ -7,12 +7,14 @@ import { ListFacade } from './+state/facade';
 import { ListReducerManager } from './+state/state.service';
 import { DataAccessCommonListModuleConfig, LIST_SERVICE } from './models/model';
 
+export const USERS_LIST_FACADE = new InjectionToken<string>('USERS_LIST_FACADE');
+
 @NgModule({
   imports: [SharedUtilNotificationModule, EffectsModule.forFeature([ListEffects])],
   providers: [ListReducerManager, ListFacade]
 })
 export class DemoDataAccessCommonListModule {
-  static config<T, S = T>(
+  static config<T extends { [key: string]: any }, S = T>(
     service: Type<ListService<T, S>>,
     config: DataAccessCommonListModuleConfig
   ): ModuleWithProviders<DemoDataAccessCommonListModule> {
